rename file name from profile to UserProfile and fix input fields issue
